Migrate add coauthors script
